// frontend/src/api/generated/models/suggestion-status-json.ts
/* tslint:disable */
/* eslint-disable */
/**
 * linkweave-api API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const SuggestionStatusJson = {
    Ok: 'OK',
    Empty: 'EMPTY',
    Unavailable: 'UNAVAILABLE',
    Preparing: 'PREPARING',
    Disabled: 'DISABLED'
} as const;
export type SuggestionStatusJson = typeof SuggestionStatusJson[keyof typeof SuggestionStatusJson];


export function instanceOfSuggestionStatusJson(value: any): boolean {
    for (const key in SuggestionStatusJson) {
        if (Object.prototype.hasOwnProperty.call(SuggestionStatusJson, key)) {
            if (SuggestionStatusJson[key as keyof typeof SuggestionStatusJson] === value) {
                return true;
            }
        }
    }
    return false;
}

export function SuggestionStatusJsonFromJSON(json: any): SuggestionStatusJson {
    return SuggestionStatusJsonFromJSONTyped(json, false);
}

export function SuggestionStatusJsonFromJSONTyped(json: any, ignoreDiscriminator: boolean): SuggestionStatusJson {
    return json as SuggestionStatusJson;
}

export function SuggestionStatusJsonToJSON(value?: SuggestionStatusJson | null): any {
    return value as any;
}

export function SuggestionStatusJsonToJSONTyped(value: any, ignoreDiscriminator: boolean): SuggestionStatusJson {
    return value as SuggestionStatusJson;
}