// frontend/src/api/generated/models/suggested-tags-result-json.ts
/* tslint:disable */
/* eslint-disable */
/**
 * linkweave-api API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TagJson } from './tag-json';
import {
    TagJsonFromJSON,
    TagJsonFromJSONTyped,
    TagJsonToJSON,
    TagJsonToJSONTyped,
} from './tag-json';
import type { SuggestionStatusJson } from './suggestion-status-json';
import {
    SuggestionStatusJsonFromJSON,
    SuggestionStatusJsonFromJSONTyped,
    SuggestionStatusJsonToJSON,
    SuggestionStatusJsonToJSONTyped,
} from './suggestion-status-json';

/**
 * 
 * @export
 * @interface SuggestedTagsResultJson
 */
export interface SuggestedTagsResultJson {
    /**
     * 
     * @type {Array<TagJson>}
     * @memberof SuggestedTagsResultJson
     */
    tagList: Array<TagJson>;
    /**
     * 
     * @type {SuggestionStatusJson}
     * @memberof SuggestedTagsResultJson
     */
    status: SuggestionStatusJson;
}



/**
 * Check if a given object implements the SuggestedTagsResultJson interface.
 */
export function instanceOfSuggestedTagsResultJson(value: object): value is SuggestedTagsResultJson {
    if (!('tagList' in value) || value['tagList'] === undefined) return false;
    if (!('status' in value) || value['status'] === undefined) return false;
    return true;
}

export function SuggestedTagsResultJsonFromJSON(json: any): SuggestedTagsResultJson {
    return SuggestedTagsResultJsonFromJSONTyped(json, false);
}

export function SuggestedTagsResultJsonFromJSONTyped(json: any, ignoreDiscriminator: boolean): SuggestedTagsResultJson {
    if (json == null) {
        return json;
    }
    return {
        
        'tagList': ((json['tagList'] as Array<any>).map(TagJsonFromJSON)),
        'status': SuggestionStatusJsonFromJSON(json['status']),
    };
}

export function SuggestedTagsResultJsonToJSON(json: any): SuggestedTagsResultJson {
    return SuggestedTagsResultJsonToJSONTyped(json, false);
}

export function SuggestedTagsResultJsonToJSONTyped(value?: SuggestedTagsResultJson | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'tagList': ((value['tagList'] as Array<any>).map(TagJsonToJSON)),
        'status': SuggestionStatusJsonToJSON(value['status']),
    };
}